requirements.txt
Django==6.0.4
psycopg[binary]>=3.2
uvicorn>=0.30
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.9.0
whitenoise>=6.6.0
Pillow
